+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 1. [](#bugfix)
     * Fixed port issue with `system.custom_base_url`
     * Hide errors with `exif_read_data` in `ImageFile`
-    * Fixed unserialize in `MarkdownFormatter` class
+    * Fixed unserialize in `MarkdownFormatter` and `Framework\File` classes
 
 # v1.7.0-rc.20
 ## 12/15/2020